+++ contrib/code_check.py
@@ -34,7 +34,7 @@ def execute(cmd: str, cmd_output: bool = False):
 
 if __name__ == "__main__":
     if not os.getcwd().endswith("weni-integrations-engine"):
-        raise Exception("The command need be executed in weni-integrations-engine")
+        raise Exception("The command need be executed in weni-marketplace-engine")
 
     # Lint validations
     execute("flake8 marketplace/")