+++ marketplace/settings.py
@@ -48,6 +48,7 @@ ALLOWED_HOSTS = env.list("ALLOWED_HOSTS")
 INSTALLED_APPS = [
     "django.contrib.admin",
     "django.contrib.auth",
+    "mozilla_django_oidc",
     "django.contrib.contenttypes",
     "django.contrib.sessions",
     "django.contrib.messages",
@@ -60,9 +61,9 @@ INSTALLED_APPS = [
     "marketplace.grpc",
     # installed apps
     "rest_framework",
-    "mozilla_django_oidc",
     "storages",
     "corsheaders",
+    "django_grpc_framework",
 ]
 
 MIDDLEWARE = [
@@ -203,7 +204,7 @@ if USE_OIDC:
 # django-cors-headers Configurations
 
 CORS_ALLOWED_ORIGINS = env.list("CORS_ALLOWED_ORIGINS", default="")
-CORS_ALLOW_ALL_ORIGINS = DEBUG
+CORS_ALLOW_ALL_ORIGINS = env.str("CORS_ALLOW_ALL_ORIGINS", default=DEBUG)
 
 
 # gRPC Connect Client
@@ -211,8 +212,12 @@ CORS_ALLOW_ALL_ORIGINS = DEBUG
 CONNECT_GRPC_SERVER_URL = env.str("CONNECT_GRPC_SERVER_URL")
 CONNECT_CERTIFICATE_GRPC_CRT = env.str("CONNECT_CERTIFICATE_GRPC_CRT", None)
 
+SOCKET_BASE_URL = env.str("SOCKET_BASE_URL", "")
+FLOWS_HOST_URL = env.str("FLOWS_HOST_URL", "")
+
 
 # Celery
+
 CELERY_BROKER_URL = env.str("CELERY_BROKER_URL", default="redis://localhost:6379")
 CELERY_RESULT_BACKEND = env.str("CELERY_RESULT_BACKEND", default="redis://localhost:6379")
 CELERY_ACCEPT_CONTENT = ["application/json"]
@@ -225,6 +230,7 @@ CELERY_TIMEZONE = TIME_ZONE
 
 APPTYPES_CLASSES = [
     "channels.weni_web_chat.type.WeniWebChatType",
+    "channels.telegram.type.TelegramType",
     "channels.whatsapp_demo.type.WhatsAppDemoType",
 ]
 
@@ -240,3 +246,12 @@ if USE_SENTRY:
         dsn=env.str("SENTRY_DSN"),
         integrations=[DjangoIntegration()],
     )
+
+
+# gRPC Framework configurations
+
+GRPC_FRAMEWORK = {
+    "ROOT_HANDLERS_HOOK": "marketplace.grpc.urls.grpc_handlers",
+}
+
+USE_GRPC = env.bool("USE_GRPC", default=False)