+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+v1.1.0
+----------
+ * Returns new fields when creating a channel
+ * Add a base class named BaseAppTypeViewSet
+ * Adjust CORS_ALLOW_ALL_ORIGINS to read environment variable
+ * Handle exceptions when creating channels
+ * Add configure endpoint to Telegram type
+ * Create Weni Web Chat Channel Using Generic Endpoint
+ * Update readme with env variables and local development instructions
+ * Add Telegram Type
+
+v1.0.0
+----------
+ * Add release channel endpoint to connect client
+ * Add project_uuid on create Weni Web Chat
+ * Rename avatarImage to profileAvatar and add openLauncherImage
+ * Add FLOWS_HOST_URL environment variable and return on configure
+ * Adjust first_name and last_name on create a new user
+ * Increase the max_length of the photo_url field to 255
+ * Add first_name and last_name in user on creating it
+ * Add sockertUrl to Weni Web Chat configure endponint
+ * Adjust featureds endpoint return
+ * Adjust customCss and channelUUID
+ * Call the Connect endpoint to create a channel in Flows
+ * Return the owner of comment in comments list endpoint 
+
+v0.1.8
+----------
+ * Add weni-protobuffers==1.1.0 package
+ * Adjust Weni Web Chat script
+ * Adjust AppTypeAsset __str__ return
+ * Add user update gRPC endpoint
+ * Adjust django admin interface static files
+
+v0.1.7
+----------
+ * Change integrations admin interface URL to `/`
+ * Adjust script that recognizes apptypes
+ * Permissions system
+
 v0.1.6
 ----------
  * Add permissions to comments endpoint